Correction to diodes.md
Description places the diodes in a Reverse Bias (Cathode on the High side, Anode on the Low side) which would prevent current flow even when the switch is pressed.

Columns must be pulled high and rows low to match the diagram and work. Looping columns vs rows first is immaterial, although it could still be rewritten to loop rows first as long as columns are the ones pulled high and not the rows.